properties in `CALL db.index.fulltext` (#1)
The property is `analyzer` and not `analyzers`. When you use `analyzers` the index is created but I think the default analyzer is used. There is no error or warning that you used a wrong/unknown property. Also, in Neo4j 3.5 you do not see what analyzer you used in `call db.indexes`.